refactor: separate business logic

+++ java-bowling/src/main/java/qna/domain/Answer.java
@@ -1,5 +1,6 @@
 package qna.domain;
 
+import qna.CannotDeleteException;
 import qna.NotFoundException;
 import qna.UnAuthorizedException;
 
@@ -44,9 +45,16 @@ public class Answer extends AbstractEntity {
         this.contents = contents;
     }
 
-    public DeleteHistory deleteAnswer() {
+    public DeleteHistory deleteAnswer(User loginUser) throws CannotDeleteException {
+        validateAnswerOwner(loginUser);
         softDeleteAnswer();
         return new DeleteHistory(ContentType.ANSWER, getId(), this.writer, LocalDateTime.now());
+    }
+
+    private void validateAnswerOwner(User loginUser) throws CannotDeleteException {
+        if (!this.writer.equals(loginUser)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
     }
 
     private void softDeleteAnswer() {
